comment out bad tests, small changes
