--- packages/core/src/private/wasm/encode-request.spec.js
@@ -1,50 +1,0 @@
-import encodeRequest from "./encode-request.js";
-
-export default test => {
-  test.case("simple encode with string body", assert => {
-    /** @type {import("#RequestFacade")} */
-    const simpleRequest =  {
-      url: new URL("https://primate.rules/my/?webserver=forever!"),
-      body: "Example string body.",
-      query: {},
-      path: {},
-      headers: {},
-      cookies: {},
-    };
-
-    const expected = [
-      0x00, 0x00, 0x00, 0x00, // url section 0
-      0x2c, 0x00, 0x00, 0x00, // length of url (44) or 0x2c
-      0x68, 0x74, 0x74, 0x70, 0x73, 0x3a, 0x2f, 0x2f, 0x70, 0x72, // 
-      0x69, 0x6d, 0x61, 0x74, 0x65, 0x2e, 0x72, 0x75, 0x6c, 0x65,
-      0x73, 0x2f, 0x6d, 0x79, 0x2f, 0x3f, 0x77, 0x65, 0x62, 0x73,
-      0x65, 0x72, 0x76, 0x65, 0x72, 0x3d, 0x66, 0x6f, 0x72, 0x65,
-      0x76, 0x65, 0x72, 0x21,
-       
-      0x01, 0x00, 0x00, 0x00, // body section
-      0x01, 0x00, 0x00, 0x00, // body kind string
-      0x14, 0x00, 0x00, 0x00, // body length (20) or 0x14
-      0x45, 0x78, 0x61, 0x6d, 0x70, 0x6c, 0x65, 0x20, 0x73, 0x74,
-      0x72, 0x69, 0x6e, 0x67, 0x20, 0x62, 0x6f, 0x64, 0x79, 0x2e,
-       
-      0x02, 0x00, 0x00, 0x00, // path section
-      0x00, 0x00, 0x00, 0x00, // no elements
-       
-      0x03, 0x00, 0x00, 0x00, // query section
-      0x00, 0x00, 0x00, 0x00, // no elements
-       
-      0x04, 0x00, 0x00, 0x00, // header section
-      0x00, 0x00, 0x00, 0x00, // no elements
-       
-      0x05, 0x00, 0x00, 0x00, // cookie section
-      0x00, 0x00, 0x00, 0x00, // no elements
-    ];
-
-    const encoded = encodeRequest(simpleRequest);
-
-    assert.equal(Array.from(encoded), expected);
-
-  });
-
-
-}
